pyproject.toml
[project]
name = "dinov3-embeddings-api"
version = "0.1.0"
description = "Batch-first FastAPI service for DINOv3 image embeddings."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "transformers==4.44.2",
  "python-multipart==0.0.9",
  "pillow==10.4.0",
  "requests==2.32.3",
]

[project.optional-dependencies]
dev = [
  "pytest==8.3.2",
  "httpx==0.27.2",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
asyncio_default_fixture_loop_scope = "function"
filterwarnings = [
  "ignore:Please use `import python_multipart` instead\\.:PendingDeprecationWarning:starlette\\.formparsers",
]
